feat: added locale.getMany
