+++ lib/create-environment-api.js
@@ -696,9 +696,13 @@ export default function createEnvironmentApi({ http, httpUpload }) {
    * .catch(console.error)
    */
   function getLocales() {
-    return http
-      .get('locales')
-      .then((response) => wrapLocaleCollection(http, response.data), errorHandler)
+    const raw = this.toPlainObject()
+    return endpoints.locale
+      .getMany(http, {
+        spaceId: raw.sys.space.sys.id,
+        environmentId: raw.sys.id,
+      })
+      .then((data) => wrapLocaleCollection(http, data))
   }
 
   /**
