+++ typings/space.d.ts
@@ -10,9 +10,10 @@ import { Environment, EnvironmentProps } from './environment'
 import { Asset, AssetProps, AssetFileProp } from './asset'
 import { ContentType } from './contentType'
 import {  ContentTypeProps} from './generated/types/content-type'
+import { LocaleProps } from './generated/types/locale'
 import { EntryProps } from './generated/types/entry'
 import { Entry } from './entry'
-import { LocaleProps, Locale } from './locale'
+import { Locale } from './locale'
 import { SpaceMember } from './spaceMember'
 import { Role, RoleProps } from './role'
 import { SpaceMembershipProps, SpaceMembership } from './spaceMembership'
